Move useful abstractions to shared library
