+++ Scripts/09-Storm.py
@@ -6,34 +6,8 @@ __doc__="""
 
 import GlyphsApp
 from NaNGFGraphikshared import *
-from NaNGFAngularizzle import *
 from NaNGFNoise import *
 from NaNFilter import NaNFilter
-from Foundation import NSMakePoint
-
-def withinGlyphBlack_faster(layer, x, y):
-	if not layer.bounds:
-		return
-	definitelyOutside = NSMakePoint(layer.bounds.origin.x-1,y)
-	pt = NSMakePoint(x,y)
-	intersections = layer.calculateIntersectionsStartPoint_endPoint_decompose_(definitelyOutside, pt, True)
-	return (len(intersections) % 2) == 1
-
-def operateOnBlackAtInterval(layer, func, step_x, step_y=None):
-	if step_y is None:
-		step_y = step_x
-	b = AllPathBounds(layer)
-	ox, oy, w, h = int(b[0])+1, b[1]+1, b[2], b[3]
-	results = []
-	for y in range(oy, oy+h, step_y):
-		for x in range(ox, ox+w, step_x):
-			if not withinGlyphBlack_faster(layer, x, y):
-				continue
-			result = func(x,y,layer)
-			if result is not None:
-				results.extend(result)
-
-	return results
 
 class Storm(NaNFilter):
 	gridsize, minsize, maxsize = 25, 20, 60
