+++ Scripts/NaNGFGraphikshared.py
@@ -5,6 +5,7 @@ import random
 import copy
 import numpy as np
 import matplotlib.path as mpltPath
+from Foundation import NSMakePoint
 
 from NaNGFConfig import *
 from NaNGFAngularizzle import *
@@ -113,6 +114,29 @@ def withinGlyphBlack(x, y, glyph):
 	else:
 		return False # not
 
+def withinLayerBlack(layer, x, y):
+	if not layer.bounds:
+		return
+	definitelyOutside = NSMakePoint(layer.bounds.origin.x-1,y)
+	pt = NSMakePoint(x,y)
+	intersections = layer.calculateIntersectionsStartPoint_endPoint_decompose_(definitelyOutside, pt, True)
+	return (len(intersections) % 2) == 1
+
+def operateOnBlackAtInterval(layer, func, step_x, step_y=None):
+	if step_y is None:
+		step_y = step_x
+	b = AllPathBounds(layer)
+	ox, oy, w, h = int(b[0])+1, b[1]+1, b[2], b[3]
+	results = []
+	for y in range(oy, oy+h, step_y):
+		for x in range(ox, ox+w, step_x):
+			if not withinLayerBlack(layer, x, y):
+				continue
+			result = func(x,y,layer)
+			if result is not None:
+				results.extend(result)
+
+	return results
 
 def point_inside_polygon_faster(x,y,poly):
